添加了将自定义按键显示在手柄界面的菜单中的功能

+++ src/com/whyun/event/ButtonTouchListener.java
@@ -35,7 +35,7 @@ public class ButtonTouchListener implements View.OnTouchListener {
 	public boolean onTouch(View v, MotionEvent event) {
 		int id = event.getAction();
 		logger.info("[motion type]" + id);
-		byte keyType = handleKeys.getTypeNow();
+		int keyType = handleKeys.getTypeNow();
 		switch (id) {
 		case MotionEvent.ACTION_DOWN: {
 			logger.info("press " + buttonId);
